+++ src/services/content/content.ts
@@ -40,5 +40,6 @@ export interface Lesson {
   title: string
   type: LessonType
   need_access: boolean
+  need_lifetime_upgrade: boolean
   page_type: LessonPageType
 }